simplify the code

+++ ft_printf/ft_puthexadecimal_lower.c
@@ -6,7 +6,7 @@
 /*   By: nash <nash@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/20 02:08:37 by nash              #+#    #+#             */
-/*   Updated: 2024/05/20 02:16:45 by nash             ###   ########.fr       */
+/*   Updated: 2024/05/20 02:47:51 by nash             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,5 @@
 
 int	ft_puthexadecimal_lower(unsigned int num)
 {
-	return (ft_puthexadecimal(num, FALSE));
+	return (ft_puthexadecimal((unsigned long long)num, FALSE));
 }
